+++ scripts/python/figures/enforcement_ratio_choropleth.py
