+++ scripts/python/figures/enforcement_ratio_location_plot.py
