+++ scripts/python/figures/enforcement_ratio_model_plots.py
