+++ scripts/python/figures/time_regression_table.py
